add bull board and verified test cases

+++ src/services/deleteUrl.services.ts
@@ -1,14 +1,14 @@
 import { db } from "../db/index";
 import { AppError } from "../errors/AppError";
 
-export const deleteUrlById = async (url_id: string) => {
-  const remove_url_query = "DELETE FROM monitor WHERE id = $1";
-  const remove_url_value = [url_id];
+export const deleteUrlById = async (url_id: string, user_id: string) => {
+  const remove_url_query = "DELETE FROM monitor WHERE id = $1 and user_id = $2";
+  const remove_url_value = [url_id, user_id];
   try {
     const deletedRows = await db.query(remove_url_query, remove_url_value);
     const rows = deletedRows.rowCount;
     if (rows === null || rows === 0) {
-      throw new AppError(400, "No such url exists");
+      throw new AppError(404, "No such url exists");
     }
     return;
   } catch (error) {
